replace employee page with new structure
